+++ packages/orm/angel_orm/example/main.dart
@@ -1,8 +1,6 @@
 import 'dart:async';
 
-import 'package:angel3_model/angel3_model.dart';
 import 'package:angel3_orm/angel3_orm.dart';
-import 'package:angel3_orm/src/query.dart';
 import 'package:angel3_serialize/angel3_serialize.dart';
 import 'package:optional/optional.dart';